+++ app/Http/Controllers/As2001KecermatanKolompertanyaanController.php
@@ -10,21 +10,50 @@ use Illuminate\Http\Response;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Log;
+use App\Services\useractivitiesService;
 use App\Services\as2001_kecermatan_kolompertanyaanService;
+use App\Libraries\branding;
 use App\Libraries\jsr;
 use App\Libraries\myfunction as fun;
+use App\Libraries\session_reader as msr;
 use Exception;
 class As2001KecermatanKolompertanyaanController extends Controller {
     //
-    protected as2001_kecermatan_kolompertanyaanService $service;
-    public function __construct(as2001_kecermatan_kolompertanyaanService $service) {
+    protected as2001_kecermatan_kolompertanyaanService|null $service;
+    protected useractivitiesService|null $activity;
+    protected branding $brand;
+    protected $titlepage, $path, $domain, $unique;
+    public function __construct(
+        Request $request,
+        as2001_kecermatan_kolompertanyaanService $service,
+        branding $brand
+    ) {
+        // ?
         $this->service = $service;
+        $this->brand = $brand;
+
+        // ?
+        $this->titlepage = $this->brand->getTitlepage();
+        $this->path      = env('SESSION_PATH', '/');
+        $this->domain    = env('SESSION_DOMAIN', 'localhosthost:8000');
+        $this->unique    = fun::random('combwisp', 50);
     }
 
     #GET
     #url = '/api/kecermatan-kolompertanyaan'
     public function all(Request $request): Response|JsonResponse|String|int|null {
         try {
+            $token = msr::read($request->bearerToken());
+            $this->activity->store([
+                'id_user'    => $token['id'],
+                'ip_address' => $request->ip(),
+                'path'       => $request->path(),
+                'url'        => $request->fullUrl(),
+                'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                'event'      => 'API - '.$request->method(),
+                'deskripsi'  => 'read : membaca semua data psikotest kecermatan.',
+                'properties' => json_encode($request->all())
+            ]);
             return jsr::print([
                 'success'   => 1,
                 'pesan'     => 'Semua Data Soal Psikotest Kecermatan!!!',
@@ -49,6 +78,17 @@ class As2001KecermatanKolompertanyaanController extends Controller {
     #url = '/api/psikotest/kecermatan/pertanyaan/{id}'
     public function allForTes(Request $request, int $id): Response|JsonResponse|String|int|null {
         try {
+            $token = msr::read($request->bearerToken());
+            $this->activity->store([
+                'id_user'    => $token['id'],
+                'ip_address' => $request->ip(),
+                'path'       => $request->path(),
+                'url'        => $request->fullUrl(),
+                'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                'event'      => 'API - '.$request->method(),
+                'deskripsi'  => 'read : membaca semua data psikotes kecermatan untuk ujian',
+                'properties' => json_encode($request->all())
+            ]);
             return $this->service->allForTes($id)->toJson();
         }
         catch(Exception $err) {
@@ -70,6 +110,17 @@ class As2001KecermatanKolompertanyaanController extends Controller {
     public function get(Request $request, String|int $val): Response|JsonResponse|String|int|null {
         try {
             $data = $this->service->get($val);
+            $token = msr::read($request->bearerToken());
+            $this->activity->store([
+                'id_user'    => $token['id'],
+                'ip_address' => $request->ip(),
+                'path'       => $request->path(),
+                'url'        => $request->fullUrl(),
+                'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                'event'      => 'API - '.$request->method(),
+                'deskripsi'  => 'read : membaca data pertanyaan kecermatan '.$data[0]['kolom_x'],
+                'properties' => json_encode($request->all())
+            ]);
             return jsr::print([
                 'success'   => 1,
                 'pesan'     => 'Data Pertanyaan Psikotest Kecermatan '.$data[0]['kolom_x'],
@@ -112,21 +163,36 @@ class As2001KecermatanKolompertanyaanController extends Controller {
                     'nilai_E' => $request->nilai_E,
                 ]);
                 if($data > 0) {
+                    $token = msr::read($request->bearerToken());
+                    $this->activity->store([
+                        'id_user'    => $token['id'],
+                        'ip_address' => $request->ip(),
+                        'path'       => $request->path(),
+                        'url'        => $request->fullUrl(),
+                        'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                        'event'      => 'API - '.$request->method(),
+                        'deskripsi'  => 'create and store : data pertanyaan kecermatan baru',
+                        'properties' => json_encode($request->all())
+                    ]);
                     return jsr::print([
                         'success' => 1,
                         'pesan'   => 'Berhasil Menyimpan Data Pertanyaan Psikotest Kecermatan!',
                         'data'    => $data
                     ], 'created');
                 }
-                return jsr::print([
-                    'error'   => 1,
-                    'pesan'   => 'Gagal Menyimpan Data Pertanyaan Psikotest Kecermatan!',
-                ], 'bad request');
+                else {
+                    return jsr::print([
+                        'error'   => 1,
+                        'pesan'   => 'Gagal Menyimpan Data Pertanyaan Psikotest Kecermatan!',
+                    ], 'bad request');
+                }
             }
-            return jsr::print([
-                'error'  => 1,
-                'pesan'  => 'Is Not Valid!',
-            ], 'not acceptable');
+            else {
+                return jsr::print([
+                    'error'  => 1,
+                    'pesan'  => 'Invalid Crendentials!',
+                ], 'not acceptable');
+            }
         }
         catch(Exception $err) {
             Log::channel('error-controllers')->error('Terjadi kesalahan pada As2001KecermatanKolompertanyaanController->store!', [
@@ -162,21 +228,36 @@ class As2001KecermatanKolompertanyaanController extends Controller {
                     'nilai_E' => $request->nilai_E,
                 ]);
                 if($data > 0) {
+                    $token = msr::read($request->bearerToken());
+                    $this->activity->store([
+                        'id_user'    => $token['id'],
+                        'ip_address' => $request->ip(),
+                        'path'       => $request->path(),
+                        'url'        => $request->fullUrl(),
+                        'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                        'event'      => 'API - '.$request->method(),
+                        'deskripsi'  => 'edit and update : data pertanyaan kecermatan yang sudah ada.',
+                        'properties' => json_encode($request->all())
+                    ]);
                     return jsr::print([
                         'success' => 1,
                         'pesan'   => 'Berhasil Memperbaharui Data Pertanyaan Psikotest Kecermatan!',
                         'data'    => $data
                     ], 'ok');
                 }
-                return jsr::print([
-                    'error'   => 1,
-                    'pesan'   => 'Gagal Memperbaharui Data Pertanyaan Psikotest Kecermatan!',
-                ], 'bad request');
+                else {
+                    return jsr::print([
+                        'error'   => 1,
+                        'pesan'   => 'Gagal Memperbaharui Data Pertanyaan Psikotest Kecermatan!',
+                    ], 'bad request');
+                }
             }
-            return jsr::print([
-                'error'  => 1,
-                'pesan'  => 'Is Not Valid!',
-            ], 'not acceptable');
+            else {
+                return jsr::print([
+                    'error'  => 2,
+                    'pesan'  => 'Invalid Credentials!',
+                ], 'not acceptable');
+            }
         }
         catch(Exception $err) {
             Log::channel('error-controllers')->error('Terjadi kesalahan pada As2001KecermatanKolompertanyaanController->update!', [
@@ -196,17 +277,42 @@ class As2001KecermatanKolompertanyaanController extends Controller {
     #url = '/api/kecermatan/kolompertanyaan/{id}'
     public function delete(Request $request, int $id): Response|JsonResponse|String|int|null {
         try {
-            $data = $this->service->delete($id);
-            if($data > 0) {
-                return jsr::print([
-                    'success' => 1,
-                    'pesan'   => 'Berhasil Menghapus Data Soal Pertanyaan Kecermatan!',
-                ], 'ok');
+            if($id) {
+                $data = $this->service->delete($id);
+                if($data > 0) {
+                    $properties = [
+                        'data' => $data,
+                        'delted_at' => date('Y-m-d H:i:s')
+                    ];
+                    $token = msr::read($request->bearerToken());
+                    $this->activity->store([
+                        'id_user'    => $token['id'],
+                        'ip_address' => $request->ip(),
+                        'path'       => $request->path(),
+                        'url'        => $request->fullUrl(),
+                        'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                        'event'      => 'API - '.$request->method(),
+                        'deskripsi'  => 'delete : data pertanyaan psikotes kecermatan ',
+                        'properties' => json_encode($properties)
+                    ]);
+                    return jsr::print([
+                        'success' => 1,
+                        'pesan'   => 'Berhasil Menghapus Data Soal Pertanyaan Kecermatan!',
+                    ], 'ok');
+                }
+                else {
+                    return jsr::print([
+                        'error'   => 1,
+                        'pesan'   => 'Gagal Menghapus Data Soal Pertanyaan Kecermatan!',
+                    ], 'bad request');
+                }
             }
-            return jsr::print([
-                'error'   => 1,
-                'pesan'   => 'Gagal Menghapus Data Soal Pertanyaan Kecermatan!',
-            ], 'bad request');
+            else {
+                return jsr::print([
+                    'error'  => 2,
+                    'pesan'  => 'Invalid Credentials!',
+                ], 'not acceptable');
+            }
         }
         catch(Exception $err) {
             Log::channel('error-controllers')->error('Terjadi kesalahan pada As2001KecermatanKolompertanyaanController->delete!', [
@@ -220,5 +326,14 @@ class As2001KecermatanKolompertanyaanController extends Controller {
                 'pesan' => 'Terjadi Kesalahan! Lihat Log!'
             ]);
         }
+    }
+
+    public function __destruct() {
+        $this->service   = null;
+        $this->activity  = null;
+        $this->titlepage = null;
+        $this->path      = null;
+        $this->domain    = null;
+        $this->unique    = null;
     }
 }
