+++ app/Http/Controllers/As2002KecermatanSoaljawabanController.php
@@ -10,15 +10,33 @@ use Illuminate\Http\Response;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Log;
+use App\Services\useractivitiesService;
 use App\Services\as2002_kecermatan_soaljawabanService;
+use App\Libraries\branding;
 use App\Libraries\jsr;
 use App\Libraries\myfunction as fun;
+use App\Libraries\session_reader as msr;
 use Exception;
 class As2002KecermatanSoaljawabanController extends Controller {
     //
-    protected as2002_kecermatan_soaljawabanService $service;
-    public function __construct(as2002_kecermatan_soaljawabanService $service) {
+    protected as2002_kecermatan_soaljawabanService|null $service;
+    protected useractivitiesService|null $activity;
+    protected branding $brand;
+    protected $titlepage, $path, $domain, $unique;
+    public function __construct(
+        Request $request,
+        as2002_kecermatan_soaljawabanService $service,
+        branding $brand
+    ) {
+        // ?
         $this->service = $service;
+        $this->brand = $brand;
+
+        // ?
+        $this->titlepage = $this->brand->getTitlepage();
+        $this->path      = env('SESSION_PATH', '/');
+        $this->domain    = env('SESSION_DOMAIN', 'localhosthost:8000');
+        $this->unique    = fun::random('combwisp', 50);
     }
 
     #GET
@@ -50,6 +68,17 @@ class As2002KecermatanSoaljawabanController extends Controller {
                 Cache::put('page-psikotest_kecermatanasoaljawaban-allForTes-'.$id, $this->service->all($id), 1*3*60*60); // 1 hari x 3 jam x 60 menit x 60 detik
                 $data = Cache::get('page-psikotest_kecermatanasoaljawaban-allForTes-'.$id);
             }
+            $token = msr::read($request->bearerToken());
+            $this->activity->store([
+                'id_user'    => $token['id'],
+                'ip_address' => $request->ip(),
+                'path'       => $request->path(),
+                'url'        => $request->fullUrl(),
+                'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                'event'      => 'API - '.$request->method(),
+                'deskripsi'  => 'read : data soal dan jawaban psikotes kecermatan ',
+                'properties' => json_encode($request->all())
+            ]);
             return $data;
         }
         catch(Exception $err) {
@@ -71,6 +100,17 @@ class As2002KecermatanSoaljawabanController extends Controller {
     public function allRaw(Request $request, int $id): Response|JsonResponse|String|int|null {
         try {
             $data = $this->service->all($id);
+            $token = msr::read($request->bearerToken());
+            $this->activity->store([
+                'id_user'    => $token['id'],
+                'ip_address' => $request->ip(),
+                'path'       => $request->path(),
+                'url'        => $request->fullUrl(),
+                'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                'event'      => 'API - '.$request->method(),
+                'deskripsi'  => 'read : data pertanyaan, soal dan jawaban psikotes kecermatan '.$data['pertanyaan'][0]['kolom_x'],
+                'properties' => json_encode($request->all())
+            ]);
             return jsr::print([
                 'success' => 1,
                 'pesan'   => 'Semua Data Pertanyaan, Soal dan Jawaban Psikotest Kecermatan '.$data['pertanyaan'][0]['kolom_x'],
@@ -96,6 +136,17 @@ class As2002KecermatanSoaljawabanController extends Controller {
     public function allCooked(Request $request, String|int $kolom): Response|JsonResponse|String|int|null {
         try {
             $data = $this->service->get($kolom);
+            $token = msr::read($request->bearerToken());
+            $this->activity->store([
+                'id_user'    => $token['id'],
+                'ip_address' => $request->ip(),
+                'path'       => $request->path(),
+                'url'        => $request->fullUrl(),
+                'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                'event'      => 'API - '.$request->method(),
+                'deskripsi'  => 'read : data pertanyaan, soal dan jawaban psikotes kecermatan '.$data['pertanyaan'][0]['kolom_x'],
+                'properties' => json_encode($request->all())
+            ]);
             return jsr::print([
                 'success' => 1,
                 'pesan'   => 'Data Pertanyaan, Soal dan Jawaban Psikotest Kecermatan : '.$data['pertanyaan'][0]['kolom_x'],
@@ -129,21 +180,36 @@ class As2002KecermatanSoaljawabanController extends Controller {
                     'soal_jawaban'  => $request->soal_jawaban,
                 ]);
                 if($data->isNotEmpty()) {
+                    $token = msr::read($request->bearerToken());
+                    $this->activity->store([
+                        'id_user'    => $token['id'],
+                        'ip_address' => $request->ip(),
+                        'path'       => $request->path(),
+                        'url'        => $request->fullUrl(),
+                        'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                        'event'      => 'API - '.$request->method(),
+                        'deskripsi'  => 'create and store : data soal dan jawaban psikotes kecermatan '.$data['kolom_x'].' baru',
+                        'properties' => json_encode($request->all())
+                    ]);
                     return jsr::print([
                         'success' => 1,
                         'pesan'   => 'Berhasil Menyimpan Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
                         'data'    => $data['data']
                     ], 'created');
                 }
-                return jsr::print([
-                    'error'   => 1,
-                    'pesan'   => 'Gagal Menyimpan Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
-                ], 'bad request');
+                else {
+                    return jsr::print([
+                        'error'   => 1,
+                        'pesan'   => 'Gagal Menyimpan Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
+                    ], 'bad request');
+                }
             }
-            return jsr::print([
-                'error'  => 1,
-                'pesan'  => 'Is Not Valid!',
-            ], 'not acceptable');
+            else {
+                return jsr::print([
+                    'error'  => 2,
+                    'pesan'  => 'Is Not Valid!',
+                ], 'Invalid Credentials!');
+            }
         }
         catch(Exception $err) {
             Log::channel('error-controllers')->error('Terjadi kesalahan pada As2002KecermatanSoaljawabanController->store!', [
@@ -171,22 +237,37 @@ class As2002KecermatanSoaljawabanController extends Controller {
                     'soal_jawaban' => $request->soal_jawaban,
                 ]);
                 if($data->isNotEmpty()) {
+                    $token = msr::read($request->bearerToken());
+                    $this->activity->store([
+                        'id_user'    => $token['id'],
+                        'ip_address' => $request->ip(),
+                        'path'       => $request->path(),
+                        'url'        => $request->fullUrl(),
+                        'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                        'event'      => 'API - '.$request->method(),
+                        'deskripsi'  => 'edit and update : data soal dan jawaban psikotes kecermatan '.$data['kolom_x'].' yang sudah ada.',
+                        'properties' => json_encode($request->all())
+                    ]);
                     return jsr::print([
                         'success' => 1,
                         'pesan'   => 'Berhasil Memperbaharui Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
                         'data'    => $data['data']
                     ], 'ok');
                 }
-                return jsr::print([
-                    'error' => 1,
-                    'pesan' => 'Gagal Memperbaharui Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
-                    'data'  => $data['data']
-                ], 'bad request');
+                else {
+                    return jsr::print([
+                        'error' => 1,
+                        'pesan' => 'Gagal Memperbaharui Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
+                        'data'  => $data['data']
+                    ], 'bad request');
+                }
             }
-            return jsr::print([
-                'error'  => 1,
-                'pesan'  => 'Is Not Valid!',
-            ], 'not acceptable');
+            else {
+                return jsr::print([
+                    'error'  => 2,
+                    'pesan'  => 'Invalid Credentials!',
+                ], 'not acceptable');
+            }
         }
         catch(Exception $err) {
             Log::channel('error-controllers')->error('Terjadi kesalahan pada As2002KecermatanSoaljawabanController->update!', [
@@ -206,19 +287,44 @@ class As2002KecermatanSoaljawabanController extends Controller {
     #url = '/api/kecermatan/soaljawaban/{id1}/{id2}'
     public function delete(Request $request, int $id1, int $id2): Response|JsonResponse|String|int|null {
         try {
-            $data = $this->service->delete($id1, $id2);
-            if($data->isNotEmpty()) {
-                return jsr::print([
-                    'success' => 1,
-                    'pesan'   => 'Berhasil Menghapus Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
-                    'data'    => $data['data']
-                ], 'ok');
+            if($id1 && $id2) {
+                $data = $this->service->delete($id1, $id2);
+                if($data->isNotEmpty()) {
+                    $properties = [
+                        'data'       => $data,
+                        'deleted_at' => date('Y-m-d H:i:s')
+                    ];
+                    $token = msr::read($request->bearerToken());
+                    $this->activity->store([
+                        'id_user'    => $token['id'],
+                        'ip_address' => $request->ip(),
+                        'path'       => $request->path(),
+                        'url'        => $request->fullUrl(),
+                        'page'       => $request->header()['titlepage'][0].$this->titlepage,
+                        'event'      => 'API - '.$request->method(),
+                        'deskripsi'  => 'delete : data soal dan jawaban psikotes kecermatan '.$data['kolom_x'],
+                        'properties' => json_encode($properties)
+                    ]);
+                    return jsr::print([
+                        'success' => 1,
+                        'pesan'   => 'Berhasil Menghapus Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
+                        'data'    => $data['data']
+                    ], 'ok');
+                }
+                else {
+                    return jsr::print([
+                        'error' => 1,
+                        'pesan' => 'Gagal Menghapus Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
+                        'data'  => $data['data']
+                    ], 'bad request');
+                }
             }
-            return jsr::print([
-                'error' => 1,
-                'pesan' => 'Gagal Menghapus Data Soal dan Jawaban Psikotest Kecermatan '.$data['kolom_x'],
-                'data'  => $data['data']
-            ], 'bad request');
+            else {
+                return jsr::print([
+                    'error'  => 2,
+                    'pesan'  => 'Invalid Credentials!',
+                ], 'not acceptable');
+            }
         }
         catch(Exception $err) {
             Log::channel('error-controllers')->error('Terjadi kesalahan pada As2002KecermatanSoaljawabanController->delete!', [
@@ -232,5 +338,14 @@ class As2002KecermatanSoaljawabanController extends Controller {
                 'pesan' => 'Terjadi Kesalahan! Lihat Log!'
             ]);
         }
+    }
+
+    public function __destruct() {
+        $this->service   = null;
+        $this->activity  = null;
+        $this->titlepage = null;
+        $this->path      = null;
+        $this->domain    = null;
+        $this->unique    = null;
     }
 }
